README PROTOTYPING: re-arranged main sections, added "Rationale" section.
The section arrangement puts the demo at the top (so new visitors can be
wow-ed and immediately immersed with the application of the project),
the setup-and-use in the middle, and the rationale at the bottom.

The "rationale" section elaborates on exactly why the script was made.
It first talks about the issues surrounding Linux Mint Cinnamon when
it came to failing to calibrate the display as well as forcing the
user (me) to rotate the display with keyboard shortcuts only
(granted xrandr exists but I wanted to rotate AND calibrate simultaneously).
It then talks about why the project solved those issues.